Fix -min, -max: type needs to be float, not double because it is accessed via a pointer to float
git-svn-id: http://svn.code.sf.net/p/netpbm/code/trunk@2619 9d0c8265-081b-0410-96cb-a4ca84ce46f8

+++ converter/other/fitstopnm.c
@@ -57,9 +57,9 @@
 struct CmdlineInfo {
     const char * inputFileName;
     unsigned int image;  /* zero if unspecified */
-    double max;
+    float max;
     unsigned int maxSpec;
-    double min;
+    float min;
     unsigned int minSpec;
     unsigned int scanmax;
     unsigned int printmax;
